requirements.txt
torch>=2.0.0
transformers>=4.49.0
accelerate>=0.20.0
datasets>=2.0.0
peft>=0.4.0
bitsandbytes>=0.41.0
tensorboard>=2.10.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pandas>=1.5.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
rapidfuzz>=3.0.0
pathlib2>=2.3.0 